Cache-snapshot 1 --> Activates the snapshot for cache modules
we have slight problem with partial store snapshots --> large memory regions results in exceeding the overhead
The overhead is placed since the file becomes to large.
If we figure out a way to store and process an sparce matrix, we can fix this problem.

Two macro added in plotter.c:
THREED 0 --> deactivates the 3D view
KEEP_TEMP 1 --> Keeps the temp file